+++ src/adminDashBoard/pages/ClientsManagement.jsx
@@ -143,8 +143,11 @@ export default function ClientsManagement() {
 
   const handleDeleteClick = async clientId => {
     setLoading(true)
-    await deleteDoc(doc(db, 'users', clientId))
+    if (window.confirm('Are you sure you want to delete this user?') === true) {
+      await deleteDoc(doc(db, 'users', clientId))
+    }
     setLoading(false)
+    return
   }
 
   const search = datas => {
