+++ src/adminDashBoard/pages/PartnersManagement.jsx
@@ -159,8 +159,11 @@ export default function PartnersManagement() {
 
   const handleDeleteClick = async clientId => {
     setLoading(true)
-    await deleteDoc(doc(db, 'users', clientId))
+    if (window.confirm('Are you sure you want to delete this user?') === true) {
+      await deleteDoc(doc(db, 'users', clientId))
+    }
     setLoading(false)
+    return
   }
 
   const getPartners = async () => {
